requirements.txt
bson==0.5.8
google==2.0.2
allure-pytest==2.5.5
allure-python-commons==2.5.5
pytest==4.4.0
python-dateutil==2.7.3
pytz==2018.4
py==1.5.4
pycparser==2.18
requests==2.19.1
requests-cache==0.4.13
ntplib==0.3.3
google-cloud-logging==1.11.0
pymongo==3.9.0
uuid==1.30
elasticsearch==7.0.3
bs4==0.0.1
html5lib==1.0.1
